standardize types across the entire client to follow via @/hooks/types.ts
